requirements.txt
Django==1.6.1
FeinCMS==1.9.1
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.3.0
South==0.8.4
boto==2.22.0
dj-database-url==0.2.2
django-cal==0.2.1
django-debug-toolbar==1.0
django-guardian==1.1.1
django-indexer==0.3.0
django-model-utils==1.5.0
django-mptt==0.6.0
django-paging==0.2.5
django-sentry==1.13.5
django-ses==0.4.1
django-storages==1.1.8
django-templatetag-sugar==0.1
docutils==0.11
eventlet==0.14.0
feedparser==5.1.3
greenlet==0.4.1
psycopg2==2.5.1
python-dateutil==1.5
raven==4.0.2
sqlparse==0.1.10
vobject==0.6.6
